Deployment/requirements.txt
Pillow==9.1.0
streamlit==1.26.0
pytube==15.0.0
moviepy==1.0.3
SpeechRecognition==3.10.0
pyvi==0.1.1
pydub==0.25.1
transformers==4.32.1